refactor: scorers
