+++ scheduling/plugins/scorers/generic.py
@@ -12,33 +12,46 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import Dict, Mapping
-from .types import Endpoint, CycleState, LLMRequest
-from .registry import register_scorer
+from __future__ import annotations
+from typing import Any, Dict, Sequence, Mapping
 import threading
+from ...framework import (
+    ScorerPlugin, 
+    Endpoint, 
+    LLMRequest, 
+    register_scorer
+)
+
+
+@register_scorer("constant")
+class ConstantScorer(ScorerPlugin):
+    """Scorer that returns a constant score."""
+
+    def __init__(self, value: float) -> None:
+        self.value = value
+
+    def score(self, cycle_state: Any, request: LLMRequest, endpoints: Sequence[Endpoint]) -> Dict[str, float]:
+        if isinstance(endpoints, Mapping):
+            return {name: float(self.value) for name in endpoints.keys()}
+        return {p.name: float(self.value) for p in endpoints}
+
 
 @register_scorer("round_robin")
-class RoundRobinScorer:
-    """A scorer that cycles through endpoints in a round-robin fashion (NeMO equivalent).
-    
-    Internal counter that increments with each score call. The endpoint corresponding to 
-    (counter % num_endpoints) receives a score of 1.0 to comply with weighted scoring.
-    """
+class RoundRobinScorer(ScorerPlugin):
+    """A scorer that cycles through endpoints in a round-robin fashion."""
 
     def __init__(self) -> None:
         self._counter = 0
         self._lock = threading.Lock()
 
-    def score(self, cycle_state: CycleState, request: LLMRequest, pods: Mapping[str, Endpoint]) -> Dict[str, float]:
+    def score(self, cycle_state: Any, request: LLMRequest, pods: Mapping[str, Endpoint]) -> Dict[str, float]:
         if not pods:
             return {}
 
-        # sorted keys ensure the idx of each pod is deterministic
         names = sorted(pods.keys(), key=str)
         with self._lock:
             idx = self._counter % len(names)
             self._counter += 1
         selected_name = names[idx]
 
-        print(f"RoundRobinScorer: selected {selected_name} (index {idx})")
         return {selected_name: 1.0}
